refactor(dir): move files
